backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dispatch"
version = "0.1.0"
description = "dispatch backend scaffold"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy[asyncio]>=2.0.31,<3.0.0",
  "asyncpg>=0.29.0,<1.0.0",
  "psycopg[binary]>=3.2.1,<4.0.0",
  "alembic>=1.13.2,<2.0.0",
  "celery[redis]>=5.4.0,<6.0.0",
  "redis>=5.0.7,<6.0.0",
  "pydantic-settings>=2.3.4,<3.0.0",
  "dnspython>=2.6.1,<3.0.0",
  "argon2-cffi>=23.1.0,<24.0.0",
  "pyotp>=2.9.0,<3.0.0",
  "itsdangerous>=2.2.0,<3.0.0",
  "cryptography>=42.0.0,<43.0.0",
  "structlog>=24.4.0,<25.0.0",
  "python-multipart>=0.0.9,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "pytest-asyncio>=0.23.8,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "ruff>=0.5.7,<1.0.0",
  "mypy>=1.11.1,<2.0.0",
  "coverage[toml]>=7.6.1,<8.0.0",
]

[tool.setuptools.packages.find]
include = ["apps*", "libs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
  "integration: integration test that may require external services",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
packages = ["apps", "libs", "tests"]
exclude = ["migrations"]

[[tool.mypy.overrides]]
module = ["celery", "celery.*"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["apps", "libs"]

[tool.coverage.report]
fail_under = 80
